+++ src/classic/index.ts
@@ -1,4 +1,4 @@
 export type { ComposedProps } from './type';
 
-export * from './injectDataWithInstance';
-export * from './injectInstanceWithContext';
+export * from './inject-data-with-instance';
+export * from './inject-instance-with-context';
