+++ src/classic/inject-data-with-instance.ts
