+++ src/classic/inject-instance-with-context.ts
